fix: use double for roadsegment positions in V2 + round on cm

+++ src/RoadRegistry/RoadSegment/Events/V1/ValueObjects/RoadSegmentLaneAttributes.cs
@@ -6,6 +6,6 @@ public class RoadSegmentLaneAttributes
     public required int AttributeId { get; set; }
     public required int Count { get; set; }
     public required string Direction { get; set; }
-    public required decimal FromPosition { get; set; }
-    public required decimal ToPosition { get; set; }
+    public required double FromPosition { get; set; }
+    public required double ToPosition { get; set; }
 }
